+++ sv-linx/src/main/java/com/hartwig/hmftools/linx/neoepitope/RefGenomeInterface.java
@@ -2,5 +2,5 @@ package com.hartwig.hmftools.linx.neoepitope;
 
 public interface RefGenomeInterface
 {
-    String getBaseString(final String chromosome, long posStart, long posEnd);
+    String getBaseString(final String chromosome, int posStart, int posEnd);
 }
